Move test helpers to helpers.php
